add handle location car

+++ SourceCode/RacingBoy/app/src/main/java/greenacademy/edu/vn/racingboy/model/Position.kt
@@ -1,6 +1,6 @@
 package greenacademy.edu.vn.racingboy.model
 
 class Position(
-        var x: Int,
-        var y: Int
+        var x: Long,
+        var y: Long
 )
